Add support for FTPing XML to server
This adds an alternative output option for the generated XML feed,
allowing it to be sent out over FTPS. It is in preparation for moving
this to an AWS Lambda.

The size of the Academic Analytics feed can get pretty big, and given
the constraints of Lambda, this uses a Unix pipe to buffer the XML. I'm
leaving the existing CLI in place because it makes it easier to test out
in development.

This also drops old versions of Python from the test suite. Although the
FTP server used for testing claims to support 2.6+, the TLS support
wasn't working in 2.7. Lambda uses 3.6 anyways.