Add support for running as AWS Lambda
This adds a Lambda handler function. There are a few assumptions being
made here. This assumes the Lambda will be triggered by a CloudWatch
event and that the event will pass a JSON structure that indicates the
type of feed to generate. The function is configured through environment
variables. Env vars that contain sensitive information should be added
to an AWS Secrets object.

I'm making some guesses here as to how the event data and secrets get
passed to the Lambda. The documentation suggests it's possible to pass
JSON in both cases, but it's not entirely clear. Some adjustment may be
required after attempting to run this in AWS.