Add build script for Lambda package
This adds a build script that can be used to create a Lambda package.
The cx_Oracle package has two shared library dependencies not present in
the Lambda execution environment. We can't distribute the Oracle client
library and it can't be automatically fetched by HTTP. The approach
taken here is to put both the Oracle client library and libaio into an
S3 bucket and fetch them at build time.

Building the Lambda package will require the AWS CLI and permissions to
read the bucket with the shared libs. Given that cx_Oracle and lxml are
both installed from platform dependent wheels, the build script should
only be used from a x86_64 linux architecture.