Update Lambda publishing step
This changes the name of the Lambda deployment package to always be
`carbon.zip`. We don't currently have the infrastructure to handle
changing Lambda aliases so we'll just overwrite the existing Lambda and
let the two week delay in staging/prod cron schedules highlight any
issues. Downgrading the Lambda to an older version can be done manually
using the deployment instructions in the README.