Add missing Symplectic FTP cert
Symplectic's FTP server has a broken cert chain. After trying to get
them to fix this it seems unlikely we'll have a solution in a timely
manner. Packaging the missing cert and adding it to the verify chain
seems like the least invasive, still reasonably secure way to fix it.