Support SSL session reuse in FTP client
There are a few problems with OpenSSL and Python's FTPs implementation.
Per RFC 4217, servers can require SSL session reuse between the command
and data channel. The stdlib ftplib module does not support this. In
these cases, OpenSSL provides poor error messaging and all you will
usually see is an Errno 0. Python's ssl module was modified in 3.6 to
support session reuse by allowing an existing session to be passed to an
SSLContext's wrap_socket method. This commit modifies the stdlib's
FTP_TLS class to make use of the new session parameter. See
https://bugs.python.org/issue19500 for background.

The second problem is that the data transfer commands--in our case,
storbinary--destroy (unwrap) the SSL session after completing the
transfer. This is problematic because with the above changes, the
session is also shared with the command channel. The unwrap call seems
to only really be necessary (I think) in cases where the underlying
unencrypted socket is going to be reused. This is not the case for us.
Calling quit() on the ftp connection should still cleanly shut down the
connection. This commit also removes the unwrap call. See
https://bugs.python.org/issue10808 for background.