Create docker build tooling
Since we are switching to Fargate, this changes the build process to
create a container. The app code itself has not changed as it already
supports this mode of operation. This commit adds the necessary tooling
to create the container.

The Makefile has been updated to handle the whole build process. Running
`make dist` should result in built container. Versioning of the app has
been switched from semver to the commit hash.

The travis deploy step will need to be updated once we are set up with
ECR.