Add Fargate functionality
This adds a few pieces to make carbon work on Fargate. The main change
is to add SNS notification directly into the app. Cloudwatch logging
from the container's stdout is a little flakey. The build process will
now publish the container to both the stage and prod ECR repos. With
real CI/CD we'd just publish to stage, but with Travis this is the best
we can do.