Add check for existence of arrive date
We ran into a case where an individual does not have a start date which
caused the upload to fail. My documentation tells me that this field is
not required for the Elements feed, so the solution here is to just omit
the field when the data is not available.