Skip people without an arrive date
If someone is missing a start date we need to skip them altogether. This
backs out the last commit and filters out people at the query level.

We have only encountered a single instance of someone missing a start
date so my assumption is that this is just a data error in some upstream
system.