Add faculty start date to people feed
Why these changes are being introduced:
For faculty, sometimes the original hire date doesn't line up with the
faculty start date and that was causing problems in the OA workflow. A
new fields was added to the data warehouse table that gives us the
date a hire became faculty, which we want to use preferentially over the
original hire date when present.

How this addresses that need:
* Adds DATE_TO_FACULTY field to the persons model
* Retrieves DATE_TO_FACULTY field from the data warehouse as part of the
  person record
* Adds a method to map the DATE_TO_FACULTY field as ArriveDate in the
  people feed XML if present, otherwise continues as before and maps the
  ORIGINAL_HIRE_DATE to ArriveDate
* Updates tests/fixtures as needed

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/INFRA-299