Preparation for AWS Organization
Why these changes are being introduced:
The Carbon app and associated infrastructure needs to be migrated to
our AWS Organization from our legacy AWS account. This will be done
with little-to-no refactoring of the application itself. But, the
repo does need some updates to match the automated deployment pipeline
in our AWS Org.

How this addresses that need:
* Update Makefile with outputs from mitlib-tf-workloads-ecr for dev
* Modify (slightly) the Makefile to adjust for S3 bucketname in Dev1
* Update Pipfile to add the "name" argument for the pypi.org source
* Add dev-build.yml workflow (from the mitlib-tf-workloads-ecr repo)
* Modify (slightly) the dev-build workflow to address the extra steps
needed in the build process that aren't currently covered by our
shared workflow

Side effects of this change:
None.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IN-568