Fix for Oracle ZIP Location
Why these changes are being introduced:
The previous Makefile had a hardcoded name for an S3 bucket that holds
a copy of the Oracle library .zip file. This makes that S3 bucket name
dynamic to match the dev/stage/prod account from which it should be
pulled.

How this addresses that need:
* Update the environment variable definitions in the Makefile

Side effects of this change:

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IN-615