IN-923 Raise exceptions for failed connection tests
Why these changes are being introduced:
* Raising exceptions for any failed connection tests stops the application
from proceeding with the workflow once a connection test fails.
This logic makes sense as the (1) a successful connection to the Data Warehouse
is required for any Carbon run and (2) a successful connection
to the Elements FTP server is required for any run Carbon run that involves FTP.

How this addresses that need:
* Exceptions are raised in the connection test functions of
carbon.database.DatabaseEngine and
carbon.app.DatabaseToFtpPipe.
* CLI tests for failed connection tests are properly configured.

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IN-923