Support incremental articles run based on env var
Why these changes are being introduced:

Previously the articles feed logic was producing an "all time"
XML file with no date filtering applied.  It was shared with us
that, as this file grew each run, it was presenting problems
for Symplectic.

How this addresses that need:

The articles feed will now respect a new env var
'ARTICLES_PUBLISH_DAYS_PAST' that, when applied, will limit the
articles included in the output to only those with a PUBLISH_DATE
on or after that many days ago.  So if today is 2025-07-01, and
this env var is set for 90 days, it will limit to rows where
PUBLISH_DATE is >= 2025-04-01, roughly speaking.

This will allow our weekly and/or monthly runs of Carbon to set
a threshold like 90 days.  Each run will be significantly smaller,
likely around 5-10mb instead of 1gb.  There will be some overlap
in the XML file between runs, but that is just fine and was the
case prior.

Side effects of this change:
* The Articles XML output will be significantly smaller when
the env var is applied.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IN-1333