Update Docker image Oracle client installation
Why these changes are being introduced:

Over time, applications that install the Oracle client have
required an update to how dependency libraries are installed.
This time has come for Carbon.

How this addresses that need:

The library `libaio1t64` is installed and symlinked for use.

The same approach was used in hrqb-client and alma-bursartransfer.

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IN-1573